use ldd to check first
